Consolidate duplicate JetpackFooter components into one
Replaces the unused component at client/components/jetpack/jetpack-footer/
with the simplified version introduced in this branch, and removes the
duplicate at client/components/jetpack-footer/. Updates all imports
(activity log, earn, podcasting, traffic) to use the canonical path.
