--- client/components/jetpack-footer/README.md
@@ -1,6 +1,0 @@
-# Jetpack Footer
-
-Component that renders Jetpack Admin Footer.
-
-Component simplified from https://github.com/Automattic/jetpack/tree/trunk/projects/js-packages/components/components/jetpack-footer
-
