--- client/components/jetpack-footer/style.scss
@@ -1,26 +1,0 @@
-@import '@wordpress/theme/design-tokens.css';
-
-.jetpack-footer {
-	border-top: var(--wpds-border-width-xs, 1px) solid var(--wpds-color-stroke-surface-neutral-weak, #e0e0e0);
-	box-sizing: border-box;
-	display: flex;
-	padding: var(--wpds-dimension-padding-xl) var(--wpds-dimension-padding-2xl);
-	width: 100%;
-}
-
-.jetpack-footer__logo {
-	flex-shrink: 0;
-}
-
-.jetpack-footer__logo-text {
-	font-size: var(--wpds-font-size-md);
-	font-weight: var(--wpds-font-weight-regular);
-	font-family: var(--wpds-font-family-body);
-	line-height: var(--wpds-font-line-height-sm);
-}
-
-a.jetpack-footer__a8c {
-	svg {
-		fill: var(--wpds-color-fg-interactive-neutral-weak);
-	}
-}
